add full python3 support

+++ statusreport.py
@@ -1,4 +1,4 @@
-import extras
+import lastpy.extras as extras
 import datetime
 import json
 
@@ -25,7 +25,7 @@ else:
     print("Start scrobbling to get your expected track count")
 
 data = extras.user_yearly_tracks(USER_NAME)
-count = data['recenttracks']['@attr']['total']
+count = int(data['recenttracks']['@attr']['total'])
 print("Year Total: " + str(count))
 yearStart = datetime.datetime(2018,1,1,0,0,0)
 # Round up the day
@@ -33,7 +33,7 @@ current = datetime.datetime.now() + datetime.timedelta(days=1)
 daysPassed = (current-yearStart).days
 # Add your own yearly scrobble goal
 yearGoal = 18000
-requirement = (yearGoal/365)*daysPassed
+requirement = int((yearGoal/365)*daysPassed)
 print("Year Goal: " + str(requirement))
 if count < requirement:
     print("You are behind. Try listening more to catch back up.")
